% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseno.R
\name{asignar_m}
\alias{asignar_m}
\title{#' Title
#'
#' @param bd
#' @param grupo
#' @param tipo
#' @param peso
#' @param metodo
#'
#' @return
#' @export
#'
#' @examples
empaquetar <- function(bd, grupo, tipo, n, peso_tamaño, criterio_n = "peso"){
aux <- bd
for(i in seq_along(tipo)){}
\usage{
asignar_m(diseño, criterio, unidades_nivel, manual)
}
\description{
\if{html}{\out{<div class="sourceCode">}}\preformatted{aux <- nivel(aux, i, grupo[i], tipo[i], n[i], peso_tamaño, criterio_n = "peso")
}\if{html}{\out{</div>}}
}
\details{

}
