% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseño.R
\name{calcular_varianza_estratificada}
\alias{calcular_varianza_estratificada}
\title{Title}
\usage{
calcular_varianza_estratificada(
  base,
  variable_estrato,
  variable_estudio,
  estimador = NULL,
  n_k = NULL,
  base_n = NULL
)
}
\arguments{
\item{base_n}{}
}
\value{

}
\description{
Title
}
