% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_ine.R
\name{CartografiaINE}
\alias{CartografiaINE}
\title{Cartografía del diseño (marco electoral INE)}
\description{
Clase R6 que almacena las cartografías electorales del diseño y provee
métodos para graficar mapas de la muestra y exportar los mapas de campo.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Cartografia-initialize}{\code{Cartografia$new()}}
    \item \href{#method-Cartografia-graficar_mapa}{\code{Cartografia$graficar_mapa()}}
    \item \href{#method-Cartografia-crear_mapas}{\code{Cartografia$crear_mapas()}}
    \item \href{#method-Cartografia-clone}{\code{Cartografia$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cartografia-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Cartografia-initialize}{}}}
\subsection{\code{Cartografia$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cartografia$new(
  df_shp = NULL,
  dl_shp = NULL,
  mun_shp = NULL,
  loc_shp = NULL,
  secc_shp = NULL,
  mza_shp = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cartografia-graficar_mapa"></a>}}
\if{latex}{\out{\hypertarget{method-Cartografia-graficar_mapa}{}}}
\subsection{\code{Cartografia$graficar_mapa()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cartografia$graficar_mapa(lflt = NULL, bd, nivel)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cartografia-crear_mapas"></a>}}
\if{latex}{\out{\hypertarget{method-Cartografia-crear_mapas}{}}}
\subsection{\code{Cartografia$crear_mapas()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cartografia$crear_mapas(
  diseño,
  zoom,
  dir,
  contexto_m = NULL,
  etiquetas_unidad = NULL
)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Cartografia-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Cartografia-clone}{}}}
\subsection{\code{Cartografia$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Cartografia$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
