% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diseño.R
\name{criterio_N}
\alias{criterio_N}
\title{Title}
\usage{
criterio_N(
  base,
  nivel,
  variable_estudio = NULL,
  num,
  criterio = "unidades",
  ultimo_nivel = F
)
}
\arguments{
\item{variable_peso}{}
}
\value{

}
\description{
Title
}
\examples{

}
