% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muestreaR-package.R
\docType{package}
\name{muestreaR-package}
\alias{muestreaR}
\alias{muestreaR-package}
\title{muestreaR: diseño de muestras polietápicas para encuestas}
\description{
Herramientas para construir el marco muestral, definir un diseño polietápico
(estratos y clusters con muestreo proporcional al tamaño), extraer la muestra,
calcular cuotas de edad/sexo y exportar los insumos de campo.
}
\details{
Las dependencias que se usan SIN prefijo \code{::} dentro del paquete se declaran
aquí para que el NAMESPACE las importe explícitamente. Sin esto, varias
funciones solo corrían si el usuario tenía esos paquetes adjuntados
(p. ej. \code{library(sf)} / \code{library(tidyverse)}). El resto de dependencias se
invocan con prefijo \code{paquete::funcion()} y no necesitan importarse aquí.
}
\author{
\strong{Maintainer}: Emilio Morones \email{emorones@morant.com.mx}

Authors:
\itemize{
  \item Emilio Morones \email{emorones@morant.com.mx}
}

}
\keyword{internal}
