% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases_ine.R
\name{PoblacionINE}
\alias{PoblacionINE}
\title{PoblaciĂ³n y marco muestral (marco electoral INE)}
\description{
Clase R6 que construye el marco muestral a partir de la lista nominal del INE
y la cartografĂ­a electoral, almacena la informaciĂ³n electoral por secciĂ³n y
permite clasificar las unidades en regiones.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Poblacion-initialize}{\code{Poblacion$new()}}
    \item \href{#method-Poblacion-calcular_poblacion}{\code{Poblacion$calcular_poblacion()}}
    \item \href{#method-Poblacion-regiones}{\code{Poblacion$regiones()}}
    \item \href{#method-Poblacion-clone}{\code{Poblacion$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poblacion-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Poblacion-initialize}{}}}
\subsection{\code{Poblacion$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Poblacion$new(nombre, ln, electoral, shp_mza, shp_loc, shp_mun)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poblacion-calcular_poblacion"></a>}}
\if{latex}{\out{\hypertarget{method-Poblacion-calcular_poblacion}{}}}
\subsection{\code{Poblacion$calcular_poblacion()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Poblacion$calcular_poblacion(na.rm = T)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poblacion-regiones"></a>}}
\if{latex}{\out{\hypertarget{method-Poblacion-regiones}{}}}
\subsection{\code{Poblacion$regiones()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Poblacion$regiones(id, regiones)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poblacion-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Poblacion-clone}{}}}
\subsection{\code{Poblacion$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Poblacion$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
