% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sustituir_muestra.R
\name{sustituir_muestra}
\alias{sustituir_muestra}
\title{Sustituir una unidad de la muestra (marco censal INEGI)}
\usage{
sustituir_muestra(diseño, shp, id, zoom, dir)
}
\arguments{
\item{diseño}{Objeto de la clase \link{Diseño} con la muestra extraída.}

\item{shp}{Lista de cartografías del diseño.}

\item{id}{Identificador de la unidad a sustituir.}

\item{zoom}{Nivel de zoom para el mapa de la unidad sustituta.}

\item{dir}{Carpeta donde se guardan los mapas.}
}
\value{
El diseño con la unidad sustituida.
}
\description{
Reemplaza una unidad seleccionada (p. ej. una manzana inviable en campo) por
otra del mismo conglomerado, recalculando la muestra y regenerando su mapa.
}
