% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dise単o.R
\name{asignar_m}
\alias{asignar_m}
\title{#' Title
#'
#' @param bd
#' @param grupo
#' @param tipo
#' @param peso
#' @param metodo
#'
#' @return
#' @export
#'
#' @examples
empaquetar <- function(bd, grupo, tipo, n, peso_tama単o, criterio_n = "peso"){
aux <- bd
for(i in seq_along(tipo)){}
\usage{
asignar_m(dise単o, criterio, unidades_nivel)
}
\description{
\preformatted{aux <- nivel(aux, i, grupo[i], tipo[i], n[i], peso_tama単o, criterio_n = "peso")
}
}
\details{

}
