% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clases.R
\name{Diseño}
\alias{Diseño}
\title{Diseño muestral polietápico (marco censal INEGI)}
\description{
Clase R6 que representa un diseño de muestra polietápico sobre el marco
censal del INEGI. Acumula los niveles (estratos y conglomerados), calcula el
plan de muestra y los factores de corrección poblacional, extrae la muestra y
calcula las cuotas de edad y sexo.
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Diseño-initialize}{\code{Diseño$new()}}
    \item \href{#method-Diseño-print}{\code{Diseño$print()}}
    \item \href{#method-Diseño-agregar_nivel}{\code{Diseño$agregar_nivel()}}
    \item \href{#method-Diseño-eliminar_nivel}{\code{Diseño$eliminar_nivel()}}
    \item \href{#method-Diseño-plan_muestra}{\code{Diseño$plan_muestra()}}
    \item \href{#method-Diseño-fpc}{\code{Diseño$fpc()}}
    \item \href{#method-Diseño-extraer_muestra}{\code{Diseño$extraer_muestra()}}
    \item \href{#method-Diseño-calcular_cuotas}{\code{Diseño$calcular_cuotas()}}
    \item \href{#method-Diseño-revisar_muestra}{\code{Diseño$revisar_muestra()}}
    \item \href{#method-Diseño-exportar}{\code{Diseño$exportar()}}
    \item \href{#method-Diseño-sustituir_muestra}{\code{Diseño$sustituir_muestra()}}
    \item \href{#method-Diseño-clone}{\code{Diseño$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-initialize}{}}}
\subsection{\code{Diseño$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$new(
  poblacion,
  n,
  n_0,
  variable_poblacional,
  unidad_muestreo,
  id_unidad_muestreo,
  llave_muestreo,
  semilla = NULL
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{semilla}}{Valor numérico opcional. Si se proporciona, el diseño es
reproducible: cada etapa estocástica fija una sub-semilla derivada de
\code{semilla}. Si es \code{NULL} (por defecto), el comportamiento es el histórico.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-print"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-print}{}}}
\subsection{\code{Diseño$print()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$print()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-agregar_nivel"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-agregar_nivel}{}}}
\subsection{\code{Diseño$agregar_nivel()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$agregar_nivel(variable, tipo, descripcion, llave)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-eliminar_nivel"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-eliminar_nivel}{}}}
\subsection{\code{Diseño$eliminar_nivel()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$eliminar_nivel(nivel)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-plan_muestra"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-plan_muestra}{}}}
\subsection{\code{Diseño$plan_muestra()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$plan_muestra(nivel, criterio, unidades_nivel)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-fpc"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-fpc}{}}}
\subsection{\code{Diseño$fpc()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$fpc(nivel)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-extraer_muestra"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-extraer_muestra}{}}}
\subsection{\code{Diseño$extraer_muestra()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$extraer_muestra(nivel)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-calcular_cuotas"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-calcular_cuotas}{}}}
\subsection{\code{Diseño$calcular_cuotas()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$calcular_cuotas()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-revisar_muestra"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-revisar_muestra}{}}}
\subsection{\code{Diseño$revisar_muestra()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$revisar_muestra(prop_vars, var_extra)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-exportar"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-exportar}{}}}
\subsection{\code{Diseño$exportar()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$exportar(shp, zoom = 16, carpeta = "Insumos", mapas = TRUE)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-sustituir_muestra"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-sustituir_muestra}{}}}
\subsection{\code{Diseño$sustituir_muestra()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$sustituir_muestra(shp, id, zoom = 16)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Diseño-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Diseño-clone}{}}}
\subsection{\code{Diseño$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Diseño$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
