% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sustituir_muestra.R
\name{sustituir_muestra_ine}
\alias{sustituir_muestra_ine}
\title{Title}
\usage{
sustituir_muestra_ine(diseño, shp, id, zoom, dir, ajustar_cuotas, crear_mapa)
}
\arguments{
\item{dir}{}
}
\description{
Title
}
